Enable support for preview features
Using Opcodes.ASM8_EXPERIMENTAL to enable analysing bytecode compiled
with --enable-preview.

The use of ASM8_EXPERIMENTAL requires the code itself using ASM to be
compiled with --enable-preview, so the library itself must be built with
enabled preview features in order to support analyzing code using
preview features. The --enable-preview flag itself requires that the
-release flag (or -source) is the same major version as the compiler
version. This means that this must be built using JDK 11. However, the
resulting library can be used with later JDKs.

As an example, this enables support for analysing code using records
introduced as preview feature in Java 14.